requirements.txt
jsonpickle==3.0.1
numpy==1.25.1
Pillow==10.0.0
pymongo==4.4.1
python-dotenv==1.0.0
ultralytics==8.0.142
